devop: add swap/tx history
