+++ package.json
@@ -60,7 +60,7 @@
     "core-js": "3.10.1",
     "graphql": "15.5.0",
     "graphql-anywhere": "4.2.7",
-    "graphql-tag": "2.11.0",
+    "graphql-tag": "2.12.4",
     "highcharts": "9.0.1",
     "highcharts-vue": "1.3.5",
     "idna-uts46": "1.1.0",
@@ -89,7 +89,7 @@
     "@vue/cli-plugin-unit-jest": "4.5.12",
     "@vue/cli-service": "4.5.12",
     "@vue/eslint-config-prettier": "6.0.0",
-    "@vue/test-utils": "1.1.4",
+    "@vue/test-utils": "1.2.0",
     "@xkeshi/vue-qrcode": "git+https://github.com/MyEtherWallet/vue-qrcode.git",
     "apollo": "2.32.5",
     "axios": "0.21.1",
@@ -137,9 +137,9 @@
     "sass-loader": "10.1.1",
     "sinon": "10.0.0",
     "store": "2.0.12",
-    "stylelint": "13.12.0",
+    "stylelint": "13.13.1",
     "stylelint-config-prettier": "8.0.2",
-    "stylelint-config-standard": "21.0.0",
+    "stylelint-config-standard": "22.0.0",
     "trezor-connect": "8.1.26",
     "u2f-api": "1.2.1",
     "uglify-es": "3.3.9",
